feat:implement controller-service-repository pattern for the back end
