+++ src/helpers/formatTranslationToOneLanguage.ts
@@ -3,7 +3,7 @@ import type { LanguageCode } from "@prisma/client";
 import type {
   findAllRestaurants,
   findRestaurantById,
-} from "~/server/api/restaurant/restaurant.service";
+} from "~/server/api/restaurant/restaurant.repository";
 
 export type RestaurantWithDetails = ReturnType<
   typeof formatTranslationToOneLanguageWithDetails
