+++ src/server/api/product/product.service.ts
@@ -1,181 +1,170 @@
-import type {
-  LanguageCode,
-  Prisma,
-  PrismaClient,
-  PrismaPromise,
-  Product,
-  ProductI18N,
-  ProductTranslationField,
-} from "@prisma/client";
+import type { ProductI18N, ProductTranslationField } from "@prisma/client";
+import { TRPCError } from "@trpc/server";
 
+import { MAX_PRODUCTS_PER_CATEGORY } from "~/config/limitations";
+import { baseErrorMessage } from "~/helpers/errorMapper";
+import * as productRepository from "~/server/api/product/product.repository";
 import type {
   CreateProductInput,
+  DeleteProductInput,
   UpdateProductInput,
   UpdateProductsPositionInput,
 } from "~/server/api/product/product.schema";
+import * as restaurantRepository from "~/server/api/restaurant/restaurant.repository";
+import type { ProtectedContext } from "~/server/api/trpc";
+import { prisma } from "~/server/db";
+import { createFieldTranslationsForAdditionalLanguages } from "~/server/helpers/createFieldTranslationsForAddtionalLanugages";
 import { formatFieldsToTranslationTable } from "~/server/helpers/formatFieldsToTranslationTable";
+import { uploadImage } from "~/server/libs/cloudinary";
 
-export const createProduct = async (
-  input: CreateProductInput & { imageUrl?: string; userId: string },
-  additionalTranslations: Pick<
+export const createProduct = async ({
+  ctx,
+  input,
+}: {
+  ctx: ProtectedContext;
+  input: CreateProductInput;
+}) => {
+  const { price, isEnabled, ...restInput } = input;
+  const userId = ctx.session.user.id;
+  let uploadedImageUrl;
+  let additionalTranslations: Pick<
     ProductI18N,
     "fieldName" | "languageCode" | "translation"
-  >[],
-  prisma: PrismaClient
-) => {
-  const { userId, price, isEnabled, ...restInput } = input;
+  >[] = [];
 
-  const translations = formatFieldsToTranslationTable<ProductTranslationField>(
-    ["name", "description"],
-    restInput
+  const restaurant = await restaurantRepository.findRestaurantByIdAndUserId(
+    { restaurantId: input.restaurantId, userId },
+    prisma
   );
 
-  const { position: biggestPosition } =
-    (await prisma.product.findFirst({
-      where: { categoryId: input.categoryId },
-      select: { position: true },
-      orderBy: { position: "desc" },
-    })) ?? {};
-
-  const nextPosition = biggestPosition ? biggestPosition + 1 : 0;
-
-  return prisma.product.create({
-    data: {
-      userId,
-      isEnabled,
-      price: price * 100,
-      imageUrl: input.imageUrl ?? "",
-      restaurantId: input.restaurantId,
-      categoryId: input.categoryId,
-      measurementUnit: input.measurementUnit,
-      measurementValue: input.measurementValue,
-      position: nextPosition,
-      productI18N: {
-        createMany: { data: [...translations, ...additionalTranslations] },
-      },
-    },
-  });
-};
-
-export const updateProduct = async (
-  input: Omit<UpdateProductInput, "isImageDeleted" | "autoTranslateEnabled"> & {
-    imageUrl?: string;
-    userId: string;
-  },
-  additionalTranslations: Pick<
-    ProductI18N,
-    "fieldName" | "languageCode" | "translation"
-  >[],
-  prisma: PrismaClient
-) => {
-  const { userId, price, isEnabled, imageUrl, ...restInput } = input;
-
-  const updatedData: Partial<Product> = {
-    price: price * 100,
-    isEnabled,
-    ...(input.measurementUnit && { measurementUnit: input.measurementUnit }),
-    ...(input.measurementValue && { measurementValue: input.measurementValue }),
-    ...(typeof imageUrl === "string" && { imageUrl }),
-  };
-
-  const translations = formatFieldsToTranslationTable<ProductTranslationField>(
-    ["name", "description"],
-    restInput
-  );
-
-  const transactions: PrismaPromise<unknown>[] = [
-    ...translations,
-    ...additionalTranslations,
-  ]
-    .filter(({ translation }) => translation)
-    .map((record) =>
-      prisma.productI18N.upsert({
-        where: {
-          productId_languageCode_fieldName: {
-            languageCode: record.languageCode,
-            fieldName: record.fieldName,
-            productId: input.productId,
-          },
-        },
-        update: {
-          translation: record.translation,
-        },
-        create: {
-          languageCode: record.languageCode,
-          fieldName: record.fieldName,
-          productId: input.productId,
-          translation: record.translation,
-        },
-      })
-    );
-
-  const [updatedProduct] = await prisma.$transaction([
-    prisma.product.update({
-      where: { id: input.productId, userId },
-      data: updatedData,
-    }),
-    ...transactions,
-  ]);
-
-  return updatedProduct;
-};
-
-export const updateManyProductsTranslations = (
-  translations: {
-    productId?: string;
-    languageCode: LanguageCode;
-    translation: string;
-    fieldName: ProductTranslationField;
-  }[],
-  prisma: PrismaClient
-) => {
-  const transactions: PrismaPromise<unknown>[] = translations
-    //TODO: ADD TS PLUGIN FROM MATT POCK TO UPDATE TYPE FOR AS STRING
-    .filter(({ translation }) => translation)
-    .map((record) => {
-      return prisma.productI18N.upsert({
-        where: {
-          productId_languageCode_fieldName: {
-            languageCode: record.languageCode,
-            fieldName: record.fieldName,
-            productId: record.productId as string,
-          },
-        },
-        update: {
-          translation: record.translation,
-        },
-        create: {
-          languageCode: record.languageCode,
-          fieldName: record.fieldName,
-          productId: record.productId as string,
-          translation: record.translation,
-        },
-      });
+  if (
+    restaurant.product.filter(
+      (product) => product.categoryId === input.categoryId
+    ).length >= MAX_PRODUCTS_PER_CATEGORY
+  ) {
+    throw new TRPCError({
+      code: "BAD_REQUEST",
+      message: baseErrorMessage.ReachedProductsLimit,
     });
+  }
 
-  return transactions;
-};
+  if (restaurant.restaurantLanguage.length > 1) {
+    additionalTranslations =
+      await createFieldTranslationsForAdditionalLanguages<ProductTranslationField>(
+        {
+          sourceLanguage: input.languageCode,
+          fieldsForTranslation: [
+            ["name", input.name],
+            ["description", input.description],
+          ],
+          restaurantLanguages: restaurant.restaurantLanguage,
+        }
+      );
+  }
+  const mainTranslations =
+    formatFieldsToTranslationTable<ProductTranslationField>(
+      ["name", "description"],
+      restInput
+    );
+  const translations = [...mainTranslations, ...additionalTranslations];
 
-export const updateManyProductsPositions = async (
-  input: UpdateProductsPositionInput,
-  userId: string,
-  prisma: PrismaClient
-) => {
-  return prisma.$transaction(
-    input.map((item) =>
-      prisma.product.update({
-        data: { position: item.position },
-        where: { id: item.id, userId },
-      })
-    )
+  if (input.imageBase64) {
+    const uploadedImage = await uploadImage(input.imageBase64, userId);
+    uploadedImageUrl = uploadedImage.url;
+  }
+
+  await productRepository.createProduct(
+    { ...input, userId, imageUrl: uploadedImageUrl },
+    translations,
+    prisma
   );
 };
 
-export const deleteProduct = async (
-  where: Prisma.ProductWhereUniqueInput,
-  prisma: PrismaClient
-) => {
-  return await prisma.product.delete({
-    where,
-  });
+export const updateProduct = async ({
+  ctx,
+  input,
+}: {
+  ctx: ProtectedContext;
+  input: UpdateProductInput;
+}) => {
+  const {
+    price,
+    isImageDeleted,
+    autoTranslateEnabled,
+    isEnabled,
+    ...restInput
+  } = input;
+  const userId = ctx.session.user.id;
+  //if image is deleted we want to remove it from db, if not keep - original in db
+  let uploadedImageUrl = isImageDeleted ? "" : undefined;
+  let additionalTranslations: Pick<
+    ProductI18N,
+    "fieldName" | "languageCode" | "translation"
+  >[] = [];
+
+  const restaurant = await restaurantRepository.findRestaurantByIdAndUserId(
+    { restaurantId: input.restaurantId, userId },
+    prisma
+  );
+
+  if (restaurant.restaurantLanguage.length > 1 && autoTranslateEnabled) {
+    additionalTranslations =
+      await createFieldTranslationsForAdditionalLanguages<ProductTranslationField>(
+        {
+          sourceLanguage: input.languageCode,
+          restaurantLanguages: restaurant.restaurantLanguage,
+          fieldsForTranslation: [
+            ["name", input.name],
+            ["description", input.description],
+          ],
+        }
+      );
+  }
+  const mainTranslations =
+    formatFieldsToTranslationTable<ProductTranslationField>(
+      ["name", "description"],
+      restInput
+    );
+  const translations = [...mainTranslations, ...additionalTranslations];
+
+  if (input.imageBase64) {
+    const uploadedImage = await uploadImage(input.imageBase64, userId);
+    uploadedImageUrl = uploadedImage.url;
+  }
+
+  await productRepository.updateProduct(
+    { ...input, userId, imageUrl: uploadedImageUrl },
+    translations,
+    prisma
+  );
+};
+
+export const updateProductsPosition = async ({
+  ctx,
+  input,
+}: {
+  ctx: ProtectedContext;
+  input: UpdateProductsPositionInput;
+}) => {
+  const userId = ctx.session.user.id;
+
+  await productRepository.updateManyProductsPositions(input, userId, prisma);
+};
+
+export const deleteProduct = async ({
+  ctx,
+  input,
+}: {
+  ctx: ProtectedContext;
+  input: DeleteProductInput;
+}) => {
+  const userId = ctx.session.user.id;
+
+  await productRepository.deleteProduct(
+    {
+      id: input.productId,
+      userId,
+    },
+    prisma
+  );
 };
